Complete Form Validation Practice Project
This commit adds teh html file with all contents, the css file with all
styles that are used, and finally the js file which validates each input.

There is a slight error where if you put in the postal code before the
country that input will be considered invalid still even after the
country has been correctly inputted. Moving the contents of the postal
code event listener into a function and then running this after
validating teh country would solve this.

This form supports most the most populated countries in the world but
not all of them.